chain_queries.rs renamed to queries.rs
